Properly handle one code path in transfer().
When transfering a range that consists of the lower part of an upper
block and the upper part of a lower block, the incorrect assertion
would kick in and the code would do the wrong thing and eventually
would lead to a crash.  We need to handle this scenario correctly.