Revert "fst: allow non-numeric key type as well"
On second thought, this does not really make much sense with the current
API, as size in shift operations is specified as key_type instead of
size_t. And the impl. imposes a lot of requirements on the key as well:
all (or nearly all) relational operators, equality operator,
meaningfully defined distance and advance by size (again for shift) and
possibly others. It is best to enforce the key to be a number.

This reverts commit 91295c7798b274006dbcc6fe87e65e691e1b2e19.
This reverts commit d8689dcc341c167f7f25e87ba6ae5a87f5d7481a.