tdf#126109 calc slow when replacing string to number
Normally, the answer to repeated erase(begin()) is to walk backwards
through the array.
However, sometimes (like here), doing so will mean that we end up
inserting at the front of a different array, which means we don't gain
anything.

So, store an extra field in the mdds block, which implements a kind
of very simple approximation of a circular array.

This gives me a 50% speedup for this bug.

This is the simplest possible thing that could work.
It could probably be made a lot more sophisticated in terms of not
wasting space.