configure.ac: remove Bashisms
The `test` built-in does not define an `==` operator in POSIX shell.
Pretending that it does leads to runtime errors when `/bin/sh` is not
Bash.

Example when `/bin/sh` is Dash:
```
./configure: 6225: test: x: unexpected operator
./configure: 6239: test: x: unexpected operator
./configure: 6266: test: xyes: unexpected operator
./configure: 6273: test: xno: unexpected operator
./configure: 6290: test: xyes: unexpected operator
./configure: 6308: test: xno: unexpected operator
```

These errors do not cause `configure` to fail outright, but they prevent
correct detection of features.

See: https://bugs.gentoo.org/837869