Define own "assert" macro and use it in test code
Trying to force assert to trigger in all build types by undef'ing NDEBUG doesn't
seem to work on Windows using MSVC.  Let's define an own macro statement and use
it.  For now it's used only in the new push_back() and emplace_back() test
cases, but all assert statements should be replaced with this going forward.